requirements.txt
Django>=2.2.6,<3.0.9
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.7.5,<2.8.0
djangorestframework-simplejwt>=4.4.0,<4.5.0

flake8>=3.8.3,<3.9.0